Change the default cache name to include $UID
This makes it so users can have their own cache by default. $UID is a variable on bash and zsh that represents the absolute ID of the user running the script. This will break compatibility with sh, but as this is a zsh plugin that isn't much of an issue. $(id -u) is an alternative, but would introduce the overhead of running an external program.